ActionSheets added.
And a bunch of other crap removed, but who knows? Git took a bunch of
stuff and chucked it when I tried to discard the changes to a file that
shouldn't have been added to begin with. No going back for you! Also,
for the record: applicationPreferences doesn't really work.